Electric Vehicle Adoption in Washington State
Electric vehicles (EVs) have emerged as a promising solution to reduce greenhouse gas emissions and dependence on fossil fuels in the transportation sector. i am aiming to explore the patterns and trends in the population of Battery Electric Vehicles (BEVs) and Plug-in Hybrid Electric Vehicles (PHEVs) registered through the Washington State Department of Licensing (DOL).